requirements.txt
django
dynaconf
django-crispy-forms
pillow
qrcode
psycopg2
